fix: lint rules
