+++ lib/src/widgets/pretty_qr_data_view.dart
@@ -1,10 +1,9 @@
-// ignore_for_file: avoid-unnecessary-setstate
+import 'dart:math' as math;
 
-import 'package:flutter/foundation.dart';
-import 'package:pretty_qr_code/src/widgets/pretty_qr_error.dart';
 import 'package:qr/qr.dart';
 import 'package:meta/meta.dart';
 import 'package:flutter/widgets.dart';
+import 'package:flutter/foundation.dart';
 
 import 'package:pretty_qr_code/src/widgets/pretty_qr_view.dart';
 import 'package:pretty_qr_code/src/painting/decoration/pretty_qr_decoration.dart';
@@ -51,8 +50,8 @@ class _PrettyQrDataViewState extends State<PrettyQrDataView> {
   late QrImage qrImage;
 
   @protected
-  // Catch all types of errors
-  // ignore: no-object-declaration
+
+  // ignore: no-object-declaration, Catch all types of errors
   Object? _lastError;
 
   @protected
@@ -61,7 +60,7 @@ class _PrettyQrDataViewState extends State<PrettyQrDataView> {
   @override
   void initState() {
     super.initState();
-    prepareQrImage();
+    _prepareQrImage();
   }
 
   @override
@@ -71,14 +70,14 @@ class _PrettyQrDataViewState extends State<PrettyQrDataView> {
     super.didUpdateWidget(oldWidget);
 
     if (oldWidget.data != widget.data) {
-      prepareQrImage();
+      _prepareQrImage();
     } else if (oldWidget.errorCorrectLevel != widget.errorCorrectLevel) {
-      prepareQrImage();
+      _prepareQrImage();
     }
   }
 
-  @protected
-  void prepareQrImage() {
+  @pragma('vm:notify-debugger-on-exception')
+  void _prepareQrImage() {
     try {
       final qrCode = QrCode.fromData(
         data: widget.data,
@@ -86,10 +85,8 @@ class _PrettyQrDataViewState extends State<PrettyQrDataView> {
       );
       qrImage = QrImage(qrCode);
     } on Exception catch (error, stackTrace) {
-      setState(() {
-        _lastError = error;
-        _lastStackTrace = stackTrace;
-      });
+      _lastError = error;
+      _lastStackTrace = stackTrace;
 
       if (widget.errorBuilder != null) {
         return;
@@ -119,13 +116,65 @@ class _PrettyQrDataViewState extends State<PrettyQrDataView> {
   Widget build(BuildContext context) {
     if (_lastError != null) {
       if (widget.errorBuilder == null) {
-        return PrettyQrErrorWidget(error: _lastError!);
+        return _PrettyQrErrorWidget(error: _lastError!);
       }
       return widget.errorBuilder!(context, _lastError!, _lastStackTrace);
     }
     return PrettyQrView(
       qrImage: qrImage,
       decoration: widget.decoration,
+    );
+  }
+}
+
+class _PrettyQrErrorWidget extends StatelessWidget {
+  // ignore: no-object-declaration, Aall types of errors
+  final Object error;
+
+  const _PrettyQrErrorWidget({
+    required this.error,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return LayoutBuilder(
+      builder: (context, constraints) {
+        final dimension = math.min(
+          constraints.maxWidth,
+          constraints.maxHeight,
+        );
+
+        if (!kDebugMode) {
+          return SizedBox.square(dimension: dimension);
+        }
+
+        return SizedBox.square(
+          dimension: dimension,
+          child: Stack(
+            alignment: Alignment.center,
+            children: [
+              const Positioned.fill(
+                child: Placeholder(
+                  color: Color(0xCF8D021F),
+                ),
+              ),
+              Padding(
+                padding: const EdgeInsets.all(4.0),
+                child: FittedBox(
+                  child: Text(
+                    '$error',
+                    textAlign: TextAlign.center,
+                    textDirection: TextDirection.ltr,
+                    style: const TextStyle(
+                      shadows: [Shadow(blurRadius: 1.0)],
+                    ),
+                  ),
+                ),
+              ),
+            ],
+          ),
+        );
+      },
     );
   }
 }
